Change round timer to integer seconds in UI and logic
Updated round timer logic to use integer seconds instead of float for display and internal calculations. Adjusted UI formatting and timer handling in scripts and prefabs to reflect this change, improving clarity and consistency for players.

+++ Assets/Scripts/Game/GameConditionManager.cs
@@ -627,9 +627,9 @@ public class GameConditionManager : MonoBehaviour
     /// <summary>
     /// Obtiene el tiempo restante entre rondas
     /// </summary>
-    public float GetTiempoRestanteEntreRondas()
+    public int GetTiempoRestanteEntreRondas()
     {
-        if (roundController == null) return 0f;
+        if (roundController == null) return 0;
         return roundController.GetTiempoRestanteEntreRondas();
     }
